Ensure all sample recipes have multiple ingredient_sections
The /recipes POST handler tests manipulate the first and last
sections in the ingredient_sections array in a variety of ways.
These test groups are only functionally different if there are at least two sections,
and the simplest way to achieve this is to manipulate a well-formed sample
recipe that already has multiple sections

This commit updates the create_recipes.py script to ensure all sample
recipes have multiple ingredient_sections, which guarantees that the
chosen sample recipe will as well.