Fix property deletion bug in /recipes/:id PUT handler
In the Recipe model, makes and serves are mutually exclusive properties
If a recipe that has one of them defined is updated to use the other,
the original property should be deleted.

Without this commit, the deletion was attempted using the code:
- delete recipe.makes (or delete recipe.serves)
- recipe.save()

This has no effect. The correct code is:
- recipe.makes = undefined (or recipe.serves = undefined)
- recipe.save()