Create a fast Mock Database
We are currently using Mockgoose to replace and mock the Mongoose database
with an in-memory database. However, populating the database for every
test is still quite slow, particularly since the tests usually modify only
a very small fraction of the database.

This commit adds a faster Mock database that shares the majority of (immutable)
state between tests, so it can be easily reset