Fix incorrect processing of 'method' in /recipes/:id PUT handler
Without this commit, the error validator for recipe.method reduces over the list of steps, looking for an error.
However, it is not initialised with a value, which makes reduce a no-op when there is only one value.
This means that errors in a single step recipe (for instance, if the step is an empty string) are undetected.

This commit initialises the reduce to null, which ensures all steps are always checked