Fix minor conceputal error in /recipes/:id PUT handler
The handler confirms that all food ids in the recipe correspond to actual food records.
Without this commit, it uses Food.find() instead of Food.findOne() which has two issues:
1. it is slightly slower, since it keeps looking for values once it found one
2. it returns an array instead of a record
Since the returned value is ignored (instead, the id is saved if no record can be found),
this does not result in incorrect behaviour.

Nevertheless, this commit fixes the conceptual error.