Create tests for the /recipes/:id PUT handler
This commit uses the new bdd-enumerator module and the Mock database to create tests for the handler.

Warning: the existing recipe tests (for the /recipes/ POST handler and the /recipes/:id GET handler)
do not yet use the bdd-enumerator module nor the Mock database and the process of integrating those
tools into the /recipes/:id PUT handler has broken some of the existing tests.
This will be rectified shortly