Remove MockDatabase Record artifacts outside spies
The MockDatabase adds two functions to every record, which allows them to
simulate the update/remove operations when record.save() and record.remove() are called.
They are necessary when the record is being handled by the code under test.

However, a number of tests also compare the records to each other. In this situation, it is more correct to remove
the mock artifacts (even though they are usually present on both objects being compared).
This is especially true when comparing the record to objects returned from the code under test,
which do not have the artificats under test and, therefore, fail deep comparisons.

This commit adds an optional flag when retrieving a record, that default to removing the mock artifacts
but can retain them if desired. It also ensures that the Mongoose spies set the flag to retain the artifacts