Fix MockDatabase to return null when findOne fails
The MockDatabase throws an error when the record cannot be found using
Model.findOne(). This is based off an incorrect understanding of how
Mongoose behaves.

I observed through experimentation that Mongoose throws an error
when an invalid ID is supplied, and I assumed this indicates that the
record could not be found. In fact, this error is due to the inability
to appropriately cast the ID before searching for it in the database.
When a well-formed ID that is not in the database is supplied,
Model.findOne should return null.

This commit updates the MockDatabase to return null for unknown ids,
and provides a magic constant A_MALFORMED_RECORD_ID that can be
used to trigger an error.

Note that with this update, many of the tests that send an invalid record
id are correctly failing, because the server responds with a 500 Error
instead of a 404.