Fix missing record processing in API handlers
The /food and /recipe handlers were written with the incorrect assumption
that the Mongoose model function 'findOne' return an error when the
record cannot be found. In fact, it only returns an error when the
id cannot be cast into an ObjectID. If the record id is well formed but
does not correspond to any record, findOne returns null.

This commit updates the handlers to account for both the second case
as well as the first one.