Remove obsolete serve-local script
The serve-local script was used to run the server on port 4000,
and to add cross-origin exceptions for port 3000 (where the CMS
was running ). This was always an undesirable workaround,
which lead to unnecessary complications.

The CMS project has been updated to use relative urls and a proxy
to interact with the API server, so the serve-local script is no
longer required. Additionally, this commit removes the unused
cross-origin logic