Allow sending plaintext cookies in development
The server policy is set to ensure cookies are always sent over
https. Cookies are used to maintain a session between client and
server, which is required for a client to access protected endpoints
without authenticating every time.

Since the development client runs over http, the authentication
mechanisms cannot be observed with the current cookie policy.

This commit sets a flag in development mode, which removes the
https cookie policy.