Reset individual spies in Mock Database
The Mock database stubs individual mongoose calls associated with
models in the database. Whenever the database is reset, the spies
are cleared with sinon.restore() and then recreated. However,
this has the side effect of clearing any spies created outside of
the database.

This commit replaces the sinon.restore() call with individual calls
for each spy created by the Mock database