Protect endpoints that change the database
All requests that change the database (POST, PUT and DELETE) should
be restricted to authenticated users.

This commit introduces middleware to return 401 responses whenever
an unauthenticated user requests a database modification through
any of the /food or /recipes endpoints.

It also updates the endpoint tests to account for the authenticated an
unauthenticated cases.