Changed default port to 5000
This project and the ovecooked-cms project both run on port 3000
by default. This has not prevented both running on the same machine
(in development) because the cms start script simply chooses a
different port if 3000 is already in use.

This behaviour does not seem to work directly on Windows machines,
which means a separate environment variable needs to be specified.

For simplicity, this commit simply changes the default API port.
Apparently, it is customary for the backend to run on port 5000
and the frontend to run on port 3000.