Add Recipe model validation and associated tests
Model validation is performed by the route handler before creating
or updating any records. However, it would be more robust if the
record structure was validated by the database as well.

This commit more completely validates Recipe records at the database
and creates a suite of record validation unit tests