Direct /cms traffic to the CMS build
The `static` directory currently has a simple `index.html` file.
It is unused since the `/` route redirects to `/api`.
The intention is to replace it with the CMS build.

This commit:
- removes the unused `index.html` file
- responds to `/cms` GET requests with `static.index.html`
- redirects users to `/cms` after github authentication completes
- updates the `/api` documentation to include the `/cms` route
- Adds the static directory to the `.gitignore` file

The last change is simply so this (API) repository on the production
server does not change whenever the CMS is updated.