Allow cookies to be sent over http in production
The server allows cookies to be sent over http during development,
but insists on https for production.

However, the production deployment does not directly modify the
node server to use https. Instead, it runs on the production server
at local port 5000, and NearlyFreeSpeech (our hosting provider)
accepts only TLS connections and proxies port 443 to 5000.

In other words, we use TLS from the browser to the production
server proxy, which then forwards the data to/from the local node
web server. This is secure, but from the point of view of the node
server, it is using http. Therefore, to support authentication,
it needs to be configured to send cookies over http.

This commit makes that change. It also removes the `serve-dev`
script, which simply set the `development` environment variable
before launching the server. Now that the development and production
server behave identically, there is no need for that script.