Serve CMS files under /cms route
The server returns the CMS `index.html` at the `/cms` route
but subsequent requests (to get the static assets) are made to
the root route. This is because the CMS is built assuming that it
is at the root of the domain.

For a variety of reasons, it is preferable to build the CMS
knowing that it will be served from `/cms`. This means that all
requests will be made under the `/cms` route.

This commit updates the server to serve static files under '/cms'