Closes #10. Convert ingredient_type to integer
In the database, Quantified ingredients and FreeText ingredients
are differentiated with the `ingredient_type` discriminator,
which is set to either `'Quantified'` or `'FreeText'`.
The discriminator must be a string.

However, we do not need to use the same discriminator when sending
recipes to clients. It is cleaner if they were mapped to the
integers 0 and 1.

This commit:
- Replaces the `'Quantified'` or `'FreeText'` ingredient
 `ingredient_type` with 0 or 1 respectively whenever recipes
  are sent to the client
- Replaces the validation code to expect 0 or 1 for `ingredient_type`
  whenever recipes are received from the client
- Updates the swagger documentation accordingly