Inform server about the NFS proxy
In production, the node server runs behind a NearlyFreeSpeech.net
TLS proxy. Our server was unaware of this. It also use `rolling`
sessions which means the cookie was sent with every response,
resetting the expiration.

This was causing issues where multiple users were sharing the same
session, as explained in overcooked-cms#29

This commit:
- Informs the server about the proxy
- Configures the proxy to only send cookies over https in production
  (now that it knows it is behind a TLS proxy)
- Adds a new script to run the server in development mode
  (allosing cookies to be sent over http)

This commit closes overcooked-cms#29.