Return updated record on PUT requests
The /food and /recipe PUT handlers return a 204 response to
indicate that the record was updated successfully.
This assumes the user knows the relevant changes and can update
their own state locally.

However, the /recipe PUT handler also updates the `lastUpdated`
timestamp, which needs to be returned to the user.

For consistency, this commit changes the /food and /recipe PUT
handlers to return the updated record

Closes #14.