Simplify indirect reference to /cms directory
The cms is built with a homepage at <server>/cms. This ensures the
service worker intercepts /cms requests and ignores data requests.
However, this also means that the index.html expects all resources
to be available at /cms/<resource>.

This is accomplished on the server by redirecting /cms to the
local resource directory.

With the new /upload route, the server also acts like a generic
file server, and the routes should not have /cms in them.

This commit removes the /cms redirection. Instead the built cms
should be uploaded to <resource dir>/cms