Use shared mock database structure
The Food route tests and Recipe route tests each independently
configure a mock database with a record structure since the
Food route has not needed to interact with Recipe records,
which will change when we prevent deletion of used food items.

This commit creates a common database structure that includes
both Food and Recipe records, and uses it in both Food and Recipe
route tests.